+++ src/stats_duck_extension.cpp
@@ -10,6 +10,7 @@
 #include "rank_correlation_function.hpp"
 #include "sign_test_function.hpp"
 #include "adjust_p_function.hpp"
+#include "table_one_function.hpp"
 #include "normality_function.hpp"
 #include "anova_function.hpp"
 #include "chisq_function.hpp"
@@ -47,6 +48,9 @@ static void LoadInternal(ExtensionLoader &loader) {
 
 	// Multiple-testing corrections
 	RegisterAdjustP(loader);
+
+	// Table 1 helper
+	RegisterTableOne(loader);
 
 	// Data import
 	RegisterReadStat(loader);